refactor: split archiver filesystem into smaller files
